+++ lib/propolis/src/hw/virtio/viona.rs
@@ -686,10 +686,10 @@ pub(crate) fn check_api_version() -> Result<(), crate::api_version::Error> {
     let vers = fd.api_version()?;
 
     // viona only requires the V2 bits for now
-    let compare = viona_api::ApiVersion::V2 as u32;
+    let compare = viona_api::ApiVersion::V2;
 
     if vers < compare {
-        Err(crate::api_version::Error::Mismatch("viona", vers, compare))
+        Err(crate::api_version::Error::Mismatch("viona", vers, compare as u32))
     } else {
         Ok(())
     }
