+++ lib/propolis/src/vmm/mod.rs
@@ -19,10 +19,14 @@ pub(crate) fn check_api_version() -> Result<(), crate::api_version::Error> {
     let vers = ctl.api_version()?;
 
     // propolis only requires the bits provided by V8, currently
-    let compare = bhyve_api::ApiVersion::V8 as u32;
+    let compare = bhyve_api::ApiVersion::V8;
 
     if vers < compare {
-        return Err(crate::api_version::Error::Mismatch("vmm", vers, compare));
+        return Err(crate::api_version::Error::Mismatch(
+            "vmm",
+            vers,
+            compare as u32,
+        ));
     }
 
     Ok(())
